+++ src/stack_pop.c
@@ -6,7 +6,7 @@
 /*   By: atabarea <atabarea@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/21 11:03:42 by alex              #+#    #+#             */
-/*   Updated: 2025/04/03 13:51:26 by atabarea         ###   ########.fr       */
+/*   Updated: 2025/04/07 14:00:23 by atabarea         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 int stack_pop(t_list *stack)
 {
     int i;
-    int value;
+    long value;
 
     if (stack->top < 0)
         return (-1);
